third_party/blink/web_tests/http/tests/cookies/partitioned-cookies/ancestor-chain-bit-redirect.https.html

<!doctype html>
<head>
<meta charset="utf-8"/>
<meta name="timeout" content="long">
<meta name="help" href="https://github.com/WICG/CHIPS#chips-cookies-having-independent-partitioned-state">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/cookies/resources/testharness-helpers.js"></script>
<title>Test partitioned cookies are sent in embedded cross-site to same-site redirects</title>
</head>
<body>
<script>

promise_test(async () => {
  //Add partitioned cookie to top-level site.
  assert_equals(document.cookie, "");
  const partitionedCookie = "foo=bar";
  const partitionedCookieAttributes =
      "; Secure; Path=/; SameSite=None; Partitioned";
  const partitionedCookieLine =
      partitionedCookie + partitionedCookieAttributes;

  document.cookie = partitionedCookieLine;

  assert_true(document.cookie.includes(partitionedCookie));

  var loads = 0;
  const sameSiteHost = window.location.hostname;
  const resourceDir = "/cookies/partitioned-cookies/resources/";
  const crossSiteUrl = new URL(
    resourceDir + "ancestor-chain-empty-embed.html",
    `https://${TEST_HOST}:${window.location.port}`);
  const sameSiteUrl = new URL(
      resourceDir + "ancestor-chain-empty-embed.html",
    `https://${sameSiteHost}:${window.location.port}`);

  // Create a cross-site iframe.
  const iframe = document.createElement("iframe");
  iframe.id = "frame";
  // Add EventListener to iframe which redirects iframe using
  // redirect-and-append-cookie-header.php after the initial load.
  // On the second load, the EventListener then evaluates the query
  // parameters added by redirect-and-append-cookie-header.php
  // to determine if the expected cookie was in the http header.
  iframe.addEventListener("load", () => {
  if(loads === 0){
    loads++;

    // Initiate the redirect from cross-site to same-site through
    // redirect-and-append-cookie-header.php
    iframe.src = resourceDir+
      "redirect-and-append-cookie-header.php?location="
      + sameSiteUrl.href;
  } else {

    // Confirm that the iframe is now same-site to the top-level site.
    const iframeUrl = new URL(iframe.src);
    const iframeHost = iframeUrl.hostname;
    assert_equals(sameSiteHost, iframeHost);

    // Confirm that the cookie was in http header of the redirect request
    assert_true(iframe.contentWindow.location.href
        .includes(partitionedCookie));
    assert_true(iframe.contentDocument.cookie
        .includes(partitionedCookie));
  }});

  //Append iframe to body to initiate the first load of the iframe
  iframe.src = crossSiteUrl.href;
  document.body.appendChild(iframe);

}, "Partitioned cookies are sent in embedded cross-site to same-site redirects");

</script>
</body>
